+++ R/Germany.R
@@ -44,8 +44,7 @@ Germany <- R6::R6Class("Germany",
 
 
     #' @description Set up a table of region codes for clean data
-    #' @importFrom tibble tibble
-    #' @importFrom dplyr mutate
+    #' @importFrom dplyr tibble
     set_region_codes = function() {
       self$codes_lookup$`1` <- tibble(
         code = c(
@@ -94,8 +93,8 @@ Germany <- R6::R6Class("Germany",
           .data$date
         ) %>%
         summarise(
-          cases_new = as.numeric(sum(.data$cases_new > 0)),
-          deaths_new = as.numeric(sum(.data$deaths_new > 0))
+          cases_new = as.numeric(sum(.data$cases_new)),
+          deaths_new = as.numeric(sum(.data$deaths_new))
         ) %>%
         ungroup()
     },
@@ -116,8 +115,8 @@ Germany <- R6::R6Class("Germany",
           .data$level_2_region, .data$date
         ) %>%
         summarise(
-          cases_new = as.numeric(sum(.data$cases_new > 0)),
-          deaths_new = as.numeric(sum(.data$deaths_new > 0))
+          cases_new = as.numeric(sum(.data$cases_new)),
+          deaths_new = as.numeric(sum(.data$deaths_new))
         ) %>%
         ungroup()
     }