+++ R/JHU.R
@@ -84,7 +84,6 @@ JHU <- R6::R6Class("JHU", # rename to country name
     source_url = "https://github.com/CSSEGISandData/COVID-19/",
 
     #' @description Set up a table of region codes for clean data
-    #' @importFrom tibble tibble
     set_region_codes = function() {
       self$codes_lookup$`1` <- JHU_codes
     },