+++ R/SouthAfrica.R
@@ -41,7 +41,7 @@ SouthAfrica <- R6::R6Class("SouthAfrica",
     source_url = "https://github.com/dsfsi/covid19za",
 
     #' @description Set up a table of region codes for clean data
-    #' @importFrom tibble tibble
+    #' @importFrom dplyr tibble
     set_region_codes = function() {
       self$codes_lookup$`1` <- tibble(
         code = c(